# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation for End user service",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API App Server"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "USERS"
                ],
                "summary": "Get all users.",
                "operationId": "bb2caab318eadb3d78d950ddf09d4958",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User uuid - **AND**",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "full_name",
                        "in": "query",
                        "description": "User full_name - **AND**",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "role_id",
                        "in": "query",
                        "description": "User role - **AND**",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersCollection"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "USERS"
                ],
                "summary": "Create user. ***DONE***",
                "operationId": "a1fce870c566eb5fc0cf3178ec5d2fed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TrvSuccessResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "USERS"
                ],
                "summary": "Get user detail",
                "operationId": "910a0c7cf580e1c605648d82a4e57f46",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "USERS"
                ],
                "summary": "Update user. ***DONE***",
                "operationId": "ddfe232db1870b82493bf8bd25596d37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TrvSuccessResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateUserRequest": {
                "required": [
                    "data"
                ],
                "properties": {
                    "full_name": {
                        "description": "User full_name",
                        "type": "string",
                        "example": "Pham Van Tien",
                        "nullable": true
                    },
                    "phone": {
                        "description": "User phoen",
                        "type": "string",
                        "example": "0123456789",
                        "nullable": true
                    },
                    "birthday": {
                        "description": "User birthday",
                        "type": "string",
                        "example": "13-08-2000",
                        "nullable": true
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": "example@gmail.com",
                        "nullable": true
                    },
                    "role_id": {
                        "description": "Role user id",
                        "type": "integer",
                        "example": "1"
                    },
                    "username": {
                        "description": "User username",
                        "type": "string",
                        "example": "phien"
                    },
                    "password": {
                        "description": "User password",
                        "type": "string",
                        "example": "Pa55w0rd"
                    },
                    "sex": {
                        "description": "MALE:0, FEMALE:1",
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": "0"
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "required": [
                    "data"
                ],
                "properties": {
                    "full_name": {
                        "description": "User full_name",
                        "type": "string",
                        "example": "Pham Van Tien",
                        "nullable": true
                    },
                    "phone": {
                        "description": "User phoen",
                        "type": "string",
                        "example": "0123456789",
                        "nullable": true
                    },
                    "birthday": {
                        "description": "User birthday",
                        "type": "string",
                        "example": "13-08-2000",
                        "nullable": true
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": "example@gmail.com",
                        "nullable": true
                    },
                    "role_id": {
                        "description": "Role user id",
                        "type": "integer",
                        "example": "1",
                        "nullable": true
                    },
                    "username": {
                        "description": "User username",
                        "type": "string",
                        "example": "phien",
                        "nullable": true
                    },
                    "password": {
                        "description": "User password",
                        "type": "string",
                        "example": "Pa55w0rd",
                        "nullable": true
                    },
                    "sex": {
                        "description": "MALE:0, FEMALE:1",
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": "0",
                        "nullable": true
                    },
                    "is_block": {
                        "description": "Block user",
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ErrorResource": {
                "properties": {
                    "data": {
                        "type": "object"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/MetaResource"
                    }
                },
                "type": "object"
            },
...
            "MetaResource": {
                "properties": {
                    "message": {
                        "description": "Meta message",
                        "type": "string"
                    },
                    "errors": {
                        "description": "Meta error",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "TrvSuccessResource": {
                "properties": {
                    "data": {
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "message": {
                                "type": "string",
                                "example": "Successful operation"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "RoleResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "full_name": {
                        "type": "string",
                        "example": "Pham Van Tien",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "example": "01234567889",
                        "nullable": true
                    },
                    "birthday": {
                        "type": "string",
                        "example": "13-08-2000"
                    },
                    "sex": {
                        "description": "0:Male, 1:Female",
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": 0,
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "example": "example@gmail.com"
                    },
                    "address": {
                        "type": "string",
                        "example": "Ha Noi"
                    },
                    "account_balance": {
                        "type": "string",
                        "example": "19036101739010"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/RoleResource"
                                }
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "UsersCollection": {
                "properties": {
                    "meta": {
                        "$ref": "#/components/schemas/MetaResource"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Enter token in format: <b>Bearer token_here</b>",
                "scheme": "bearer"
            },
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "trv-api-key"
            }
        }
    }
}